Add solution for problem `E - Four Segments` Link: https://codeforces.com/contest/1468/problem/E Tags: greedy, *800 Ref: https://codeforces.com/contest/1468/submission/103610470

+++ codeforces/1468/E.cpp
@@ -37,8 +37,7 @@ int main(){
     int t;
     cin>>t;
     while(t--){
-        ll a[4];
-
+        int a[4];
         for(auto &i:a){
             cin>>i;
         }
